Add support for having warnings vs errors
A basic framework for dealing with issues that users may want to treat as either
warnings or errors.

It supports setting a default from compile time, modifying it at runtime using an
environment variable, and push/popping states off of a stack.

It also changes some exceptions that were being thrown to issues specifically because
those exceptions are making it very difficult for me to transition my codebase from
OCMock 3.4.1 to OCMock 3.6, and I know it will also be an issue for
https://github.com/TextureGroup/Texture moving to OCMock 3.6.