Avoid problems when KVO is still registered on a mock object that is being stopped.
The recommended pattern for KVO is now to let the runtime take care of unregistering
KVO observers at dealloc. This means that observers are often still registered on mocks
when they are being stopped. Both mocks and KVO change up the class tree and can easily
step on each other. We will do our best to clean up the classes we create, but "leaking"
a class is preferable to forcing users to deregister their KVO.