Retain mocks to improve thread safety
Mock objects are currently only retained by the thread that created them.
This patch retains the mock by the thread calling it for the lifetime of the call.
Previously Thread A that created the mock could kill it off while Thread B that was
calling the mock was using it.
This greatly improves the thread safety of using mocks (although it's still risky).