Xcode12: silence warning about minimum deployment
generally speaking, it will often be the case that OCMock iOS target
will only be part of a build on simulator for unit testing.

in such a case, a warning is emitted that the project settings for
the minimum deployment target (8.0) are less than what Xcode12
provides (9.0 to 14.0.99) .

by establishing an iOS 14 SDK specific value, then anyone still using
this project in Xcode 11 (or previously) will still be able to build
against iOS 8 simulators ... but those building on Xcode 12 won't be
subject to warnings.