Allow injection of macro state class into OCM macros
We have a need to inject a replacement for `OCMMacroState` for when we are doing cross
process testing using a library like https://github.com/google/EarlGrey/tree/earlgrey2.

This simple expansion to the basic macros will give us the flexibility we need to allow
us to use OCMock v3 syntax instead of falling back to OCMock v2 syntax when working
with EarlGrey tests.

The modification to andReturn macro allows us to move objects across the process boundary
instead of "hiding" an object instead of an NSValue. The "hack" of using NSArray as a
tuple could be replaced by a real object, or an NSDictionary, but I went with the simple
stupid approach to start.