Make expectation recording thread-safe.
To handle a mocked invocation, we do the following:
1. Get the first stub that matches the invocation.
2. If the stub is an expectation, remove the stub from the arrays of stubs and expectations.
3. Pass the invocation to the stub.

The first two steps are currently not atomic, so there is a race condition. For example:
1. Add expectation 1 for method A.
2. Add expectation 2 for same method A.
3. [Thread 1] Call method A.
4. [Thread 2] Call method A.
5. [Thread 1] Get expectation 1 since it matches the invocation.
6. [Thread 2] Get expectation 1 since it matches the invocation.
7. [Thread 1] Remove expectation 1 from the arrays.
8. [Thread 2] Remove expectation 1 from the arrays.

In the above example, the invocations of the same method in Thread 1 and Thread 2 both match expectation 1; expectation 2 does not get matched.

The solution is to make the matching and the removal atomic. We can do so by wrapping the relevant code in `@synchronized(stubs)`. (Even though `stubForInvocation:` also performs `@synchronized(stubs)`, we can feel free to do the same in the caller without fear of deadlocks since `@synchronized` uses a recursive lock.)

Added a new unit test that was failing before the fix and is now passing after the fix.